Handle retained_data saving when logged out more gracefully.
Before this change, if they had a tab left open to an assignment
or module and they got logged out somehow. E.g the session expired or
they logged out from a different tab or browser, the user could
continue filling out fields and it would warn that data wasn't saving
b/c they weren't connected to the internet. If the user continued filling out
fields, for each one, a retry would kickoff in the background every 5 seconds.
This lead to:

1) Users freaking out b/c they would do a ton of work and it wouldn't save even though
   they were connected to the internet
2) Our servers being hammered with calls to set_user_retained_data if several users
   got into this situation and left their tabs open. For example, in one week we saw
   1.16 million calls to this endpoint -- far surpassing what would be expected for 500
   currently enrolled students.

Note: this problem was exacerbated by a separate bug which caused you to get logged out
after 1 day of inactivity, so tabs left open for a day would start to exhibit this behavior.
That was fixed here: https://github.com/beyond-z/canvas-lms/pull/572

Test Plan:
1. Login and fill out some fields in an assignment.
2. In a separate tab, hit the Logout button.
3. In the original tab, continue filling out fields.
4. Notice that the message about being logged out is shown instructing them they're being redirected
   and then a prompt comes up telling them their data didn't save and asking if they want to leave anyway.

- Repeat the above steps but on step 4 do the following:
  - Logout in a different browser
  - Clear the cookies
  - Do a server restart => expected behavior is that they are still logged in and none of this stuff happens.

- Repeat the above steps for a module.
- Repeat the above steps with the dynamic syllabus (aka Course Home) loaded for regression purposes
  to make sure that when they click something the normal redirect behavior happens.